requirements.txt
telethon==1.37.0
requests==2.32.3
langchain-community
tiktoken
langchain-openai
langchainhub
chromadb
langchain
langgraph
python-dotenv
langchain_chroma
openai